Add files via upload
Fundamental changes in this new version.

Previous versions had the issue that each time a new panel or url was loaded a new tab was created in chrome. Over time you got more and more tabs. This slowed down the system more and more.
In the new version of the mqttDisiplayClient chrome is controlled over the DevToolsAPI of chrome.
You have control now how many tabs are created and of the lifetime of the tabs.

See the Readme.md how to enable DevTools API in chrome and how to configure chrome in the ini file.
check also the remarks how ro configure URLs for the panels in the ini file. This is important to control the tabs in chrome correctly.

Additional changes:

- Home Assistant discovery added
- All files are pylint conform. Because of this, all python files have pylint conform names now.
- The blank_page switch topic is removed. You can select a blank panel with the panel topic
- It you send a url to the url topic, the url is now imediately shown and the panal topic switches to URL
- url section in ini file removed
- chrome and haDiscover section added in ini file
- The client has now an integration test frame to verify the functionality